streamlit/requirements.txt
opencv-python-headless
face-recognition
streamlit
Pillow
pickle-mixin
